requirements.txt
absl-py==2.3.1
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
docopt==0.6.2
et_xmlfile==2.0.0
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
future==1.0.0
google-auth==2.45.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pipreqs==0.4.13
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
torch==2.4.1+cu118
torchsummary==1.5.1
torchvision==0.19.1+cu118
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.2.3
Werkzeug==3.0.6
yarg==0.1.10
zipp==3.20.2
